+++ src/Core/PayPalDefinitions.php
@@ -21,6 +21,7 @@ final class PayPalDefinitions
     public const SEPA_PAYPAL_PAYMENT_ID = 'oscpaypal_sepa';
     public const CCALTERNATIVE_PAYPAL_PAYMENT_ID = 'oscpaypal_cc_alternative';
     public const GOOGLEPAY_PAYPAL_PAYMENT_ID = 'oscpaypal_googlepay';
+    public const APPLEPAY_PAYPAL_PAYMENT_ID = 'oscpaypal_apple_pay';
 
     //vaulting
     public const PAYMENT_VAULTING = [
@@ -87,6 +88,28 @@ final class PayPalDefinitions
             'countries' => [],
             'currencies' => ['AUD', 'BRL', 'CAD', 'CNY', 'CZK', 'DKK', 'EUR', 'HKD', 'HUF', 'ILS', 'JPY', 'MYR', 'MXN', 'TWD', 'NZD', 'NOK', 'PHP', 'PLN', 'GBP', 'RUB', 'SGD', 'SEK', 'CHF', 'THB', 'USD'],
             'uapmpaymentsource' => 'googlepay',
+            'constraints' => self::PAYMENT_CONSTRAINTS_PAYPAL,
+            'onlybrutto' => false,
+            'buttonpayment' => false,
+            'defaulton' => true
+        ],
+        //ApplePay
+        self::APPLEPAY_PAYPAL_PAYMENT_ID => [
+            'descriptions' => [
+                'de' => [
+                    'desc' => 'ApplePay',
+                    'longdesc' => '',
+                    'longdesc_beta' => 'Bezahlen Sie bequem mit ApplePay. Starten Sie direkt von der Detailsseite oder im Warenkorb.'
+                ],
+                'en' => [
+                    'desc' => 'ApplePay',
+                    'longdesc' => '',
+                    'longdesc_beta' => 'Pay conveniently with ApplePay. Start directly from the details page or in the shopping cart.'
+                ]
+            ],
+            'countries' => [],
+            'currencies' => ['AUD', 'BRL', 'CAD', 'CNY', 'CZK', 'DKK', 'EUR', 'HKD', 'HUF', 'ILS', 'JPY', 'MYR', 'MXN', 'TWD', 'NZD', 'NOK', 'PHP', 'PLN', 'GBP', 'RUB', 'SGD', 'SEK', 'CHF', 'THB', 'USD'],
+            'uapmpaymentsource' => 'apple_pay',
             'constraints' => self::PAYMENT_CONSTRAINTS_PAYPAL,
             'onlybrutto' => false,
             'buttonpayment' => false,